+++ web/src/main/java/com/homeapp/javatraining/dto/mapper/UserMapper.java
@@ -1,10 +1,12 @@
 package com.homeapp.javatraining.dto.mapper;
 
 import com.homeapp.javatraining.dto.ProfileResponse;
+import com.homeapp.javatraining.dto.UserDTO;
 import com.homeapp.javatraining.model.User;
 import org.mapstruct.Mapper;
 
 @Mapper(componentModel = "spring")
 public interface UserMapper {
     ProfileResponse toProfileResponse(User user);
+    UserDTO toUserDTO(User user);
 }